pyproject.toml
[project]
name = "climatesentiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bertopic>=0.17.3",
    "datetime>=5.5",
    "nltk>=3.9.1",
    "pandas>=2.3.1",
    "praw>=7.8.1",
    "psaw>=0.1.0",
    "python-dotenv>=1.1.1",
]
